+++ internal/resource/infrastructure_stack.go
@@ -7,6 +7,7 @@ import (
 
 	"terraform-provider-plural/internal/client"
 	"terraform-provider-plural/internal/common"
+	"terraform-provider-plural/internal/model"
 
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
@@ -52,7 +53,7 @@ func (r *InfrastructureStackResource) Configure(_ context.Context, req resource.
 }
 
 func (r *InfrastructureStackResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
-	data := new(infrastructureStack)
+	data := new(model.InfrastructureStack)
 	resp.Diagnostics.Append(req.Plan.Get(ctx, data)...)
 	if resp.Diagnostics.HasError() {
 		return
@@ -74,7 +75,7 @@ func (r *InfrastructureStackResource) Create(ctx context.Context, req resource.C
 }
 
 func (r *InfrastructureStackResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
-	data := new(infrastructureStack)
+	data := new(model.InfrastructureStack)
 	resp.Diagnostics.Append(req.State.Get(ctx, data)...)
 	if resp.Diagnostics.HasError() {
 		return
@@ -91,7 +92,7 @@ func (r *InfrastructureStackResource) Read(ctx context.Context, req resource.Rea
 }
 
 func (r *InfrastructureStackResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
-	data := new(infrastructureStack)
+	data := new(model.InfrastructureStack)
 	resp.Diagnostics.Append(req.Plan.Get(ctx, data)...)
 	if resp.Diagnostics.HasError() {
 		return
@@ -112,7 +113,7 @@ func (r *InfrastructureStackResource) Update(ctx context.Context, req resource.U
 }
 
 func (r *InfrastructureStackResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
-	data := new(infrastructureStack)
+	data := new(model.InfrastructureStack)
 	resp.Diagnostics.Append(req.State.Get(ctx, data)...)
 	if resp.Diagnostics.HasError() {
 		return
